+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+[Compare the full difference](https://github.com/andrlik/django-podcast-analyzer/compare/v0.2.2...HEAD)
+
+- Add fix for cases with podcastparser doesn't split keywords correctly.
+
 ## 0.2.2
 
 [Compare the full difference](https://github.com/andrlik/django-podcast-analyzer/compare/v0.2.1...0.2.2)